+++ sdk/cosmos/azure_data_cosmos/tests/multi_region.rs
@@ -1,3 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-mod multi_region_tests;