+++ sdk/cosmos/azure_data_cosmos/tests/multi_region_tests/mod.rs
@@ -1,6 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-mod cosmos_multi_region_fallback;
-
-#[path = "../framework/mod.rs"]
-mod framework;