+++ sdk/cosmos/azure_data_cosmos_driver/tests/multi_write_tests/mod.rs
@@ -1,5 +1,4 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-pub mod driver_excluded_regions_fallback;
 pub mod driver_partition_failover;